agents/worldcoin_ask_world/requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
uagents>=0.4.0 